# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Project 1 Kampus Merdeka X Hacktiv8 API Docs",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/todos": {
            "get": {
                "description": "Get All Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get All Data Todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todos.Todo"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Edit Data Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Put Edit Data Todos",
                "parameters": [
                    {
                        "description": "Update",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/todos.CreateTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessUpdate"
                        }
                    }
                }
            },
            "post": {
                "description": "Post New Data Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Post New Data Todos",
                "parameters": [
                    {
                        "description": "Todos",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/todos.CreateTodo"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessCreate"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Data Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Data Todos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Todo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessDelete"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.SuccessCreate": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/todos.Todo"
                },
                "messages": {
                    "type": "string",
                    "example": "success add new todos"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessDelete": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "success delete todos"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessUpdate": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "success delete todos"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "todos.CreateTodo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Cuci Baju"
                }
            }
        },
        "todos.Todo": {
            "type": "object",
            "properties": {
                "complete": {
                    "type": "boolean",
                    "example": false
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Cuci Baju"
                }
            }
        }
    }
}